agent/requirements.txt
requests==2.31.0
watchdog==3.0.0
pillow==10.2.0
exifread==3.0.0
